@@ -3,7 +3,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.santiagolizardo.neocommander</groupId>
 	<artifactId>NeoCommander</artifactId>
-	<version>1.4.11</version>
+	<version>1.4.12</version>
 	<packaging>jar</packaging>
 	<name>NeoCommander</name>
 	<url>https://github.com/santiagolizardo/neocommander</url>
@@ -35,13 +35,13 @@
 				<artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <release>17</release>
+                    <release>21</release>
                     <compilerArgs>
                         <arg>-Xlint:deprecation</arg>
                         <arg>-Xlint:unchecked</arg>
                     </compilerArgs>
-                    <source>17</source>
-                    <target>17</target>
+                    <source>21</source>
+                    <target>21</target>
                 </configuration>
 			</plugin>
 			<plugin>
